started a list of jobs I've applied for
I also made a short HTML file that I could give when sites ask for a
portfolio document. All it does is redirect to the portfolio repo I
have. This is so I don't have to maintain a separate PDF... I really
should turn that repo into a github.io site.